Delayed feedback expressions now supported
Pymoca now supports modelica delay() expressions. This commit updates
ModelicaMixin to use the improved Pymoca API, as well as the changes
needed in transcribe() to resolve constraints associated with delay
expressions and durations.

So far, delay durations may only consist of expressions of parameters,
numerical values, and constant inputs so as to ensure delay durations
are resolvable to numerical values at transcribe() time.