Codecov to only check diff coverage in MRs
Currently we could get a pipeline error in the coverage stage because
the diff (changes introduced by commit) were not covered in any tests.
This is a different check than the "project" one, but we only want those
checks in MRs. As soon as we merge, even with a decrease in coverage, we
want the master pipeline to succeed.