Add console scripts for accessible examples and libraries
The examples for RTC-Tools do not come packaged in the pip package, and
neither would the typical installation location be very convenient.
Editable installs with a git repository are of course possible for
developers, but end users need something easier and more accessible.

The same goes for the Modelica libraries. They are now also included as
pip packages, and the location of the Modelica files is therefore not
very accessible.

To solve this, we add two console scripts, that will get placed in PATH
upon installation:
- rtc-tools-download-examples
- rtc-tools-copy-libraries

Note that the Modelica libraries are _copied_ and not symlinked, to
avoid confusion when editing files, and also avoid users inadvertently
deleting the actual Modelica files in the package.

Closes #1032 and #1034