Disable codecov patch coverage errors
For some reason the "only_pulls" option did not work as expected. When
merging a request, the GitLab/webhook would still fill the "pull" field
with the corresponding number, instead of leaving it empty.

The coverage report on the diff/patch is still useful, so we do not
disable the option altogether.