Fix lingering reference to GPL
Contrary to the GPL, the LGPL does not require printing a notification
to the screen. We therefore only log the version number, which could
still be useful information.

Most other references to GPL were taken care of in bc4ebe74, but this
one was missed. Note that a similar message for optimization problems
was already addressed in said commit.