Fix delayed feedback handling for multiple delays
When more than one delay variable is present, multiple calculations would
fail.

The division for history calculations would fail because of mismatching
shapes. By default, NumPy would try to divide every row, and therefore
the number of columns should be equal. In our case, we want to divide
every column by some other list of values, so we have to use [:, None] to
index it. Furthermore, the calculation has been rewritten to use np.diff for
clarity/simplicity.

The delayed_feedback_function returns a list of values, and these should
not be unpacked.

Lastly, the truth value of an array is not allowed either, and requires
either all/any to reduce it to a single boolean first.