Fix history extrapolation for delayed expressions
The supposed block interpolation that was in place does not work as
suggested. It would only properly extrapolate values if the corresponding
output x-coordinate was outside of the range of input x-coordinates. It
would however not extrapolate or interpolate like that with NaNs present.
For example (y-coordinates of input and output only):
          [4,0, 4.0] --> [4.0, 4.0, 4.0, 4.0]
[nan, nan, 4.0, 4.0] --> [nan, nan, 4.0, 4.0]