Fix unwanted conversion to MX in substitute calls
It could happen that a DM expression would be converted to an MX
expression, which would then later confuse the solver (e.g. when the
values are the lower/upper bounds of the constraints).

One example is that when passing as expr a 15x1 DM array of all 293.55
values, the resulting value would be "MX(all_293.65(15x1))". This in
turn would confuse the solver call, resulting in an exception.