Fix goal consistency and monotonicity checks
If a goal had both a target min and a target max, the latter would not
get checked for consistency due to the "elif" instead of "if".

The check for target min/max exceedence of the function range compared
the True/False value returned by has_target_* instead of the actual
target min/max value.

Furthermore, the monotonicity checks path goals were behind the option
flag, but those of (normal) goals were not.