Fix goal constraint min/max bounds consistency
It could happen that lower (later) priority goals for an already
constrained state would result in an answer that would be slightly
outside the hard constraints due to feasibilty tolerances.
When combining the soft constraint and existing hard constraint in a new
hard constraint, the existing hard constraint values could be shifted,
basically enforcing a solution that may or may not be feasible.

An example for a typical state with goals at multiple priorities:

+----------+------------+------------+
| Priority | Target Min | Target Max |
+----------+------------+------------+
|        1 |        0.0 |        0.7 |
|        2 |        0.7 |        0.7 |
+----------+------------+------------+

The first priority could end up with a hard constraint exactly matching,
i.e. (0.0, 0.7). Exaggerating the numerical differences that can occur,
the second priority could end up with a value of 0.701 for the state in
question.
What would happen before this commit is that we would calculate a new
hard constraint (0.701, 0.7). This would then be corrected to (0.701,
0.701), which would then replace the original hard constraint.

It is safer to keep enforcing the old hard constraint (which led to the
solution of 0.701) making the new constraint (0.7, 0.7). This is the
behavior that this commit makes happen.