Fix goal monotonicity checks
Goals should only be compared to other goals with the same function key,
not _all_ other goals. Also make sure to loop over all ensemble members,
which is an argument to the get_function_key() method of a goal.

Furthermore, we were comparing the target max of "other" to the target
_min_ of "goal". We should of course be comparing the target max of both
goals to each other.

Regression in 1b9cc608.