Fix big-M in goal programming example
Not sure why a very large value was chosen, as it is well established
that big-M values should typically be as small as possible. Very high
values can result in bad relaxations and make branching very hard.

In this case, a more reasonable M of 2 is both faster (~2x), and also
results in a better solution. The first two priorities remain the same,
but the third priority goes from 12.23 to 10.85.