Fix parameter inlining for path objective/constraints
When using goal programming, parameters would only get inlined in the
first priority (i.e. the first call to the transcribe() method).

Note that regular objective and constraints cannot/should not refer to
parameters with the variable() or state() methods (which return an MX
symbol), and instead always rely on state_at() and parameters(). The
latter two always return a numberic value, and we therefore we do not
need to perform any inlining on the regular objective and constraints.