Rename test models to be able to use pytest
The naming of some classes confused pytest. In general, it looks for
classes and functions starting with "Test" or "test". In our case this
meant it would also find (and fail on) the OptimizationProblem classes.
Renaming all such classes from "TestProblem*" to "Model* fixes this.

With nose in maintenance mode and pytest getting more popular, it makes
sense to make this small renaming change to support both.

Closes #1033