Fix discrepancy between goal and path goal scaling
Path goals were not divided by the number of time steps, meaning that
formulting a problem with a bunch of normal goals using state_at() (one
for each time step) and a path goal would lead to different objectives and
results.

To fix the discrepancy, we disable the division by the number of
goals/time steps by default. This is conservative in the sense that we
then guarantee at least as accurate a result as before this commit. For
very tricky/badly scaled problems it may however lead to convergence
issues.

Should scaling with the problem size be desired, the new
"scale_by_problem_size" goal programming option can be set to True.

Closes #1046