Improve goal function range vs. targets check
We now check that the target min/max are not equal to the function range
lower bound/upper bound respectively as well.

A target minimum equal to the lower bound of the function range does not
make any sense, as the corresponding epsilon is then free (and will
therefore always become zero). Effectively, the goal will behave as a
critical goal, and users should use those instead if such behavior is
desired.