Only do duplicate parameter check for PI-XML
Before, we would (optionally) check if a parameter read from the PI-XML
files (rtcParameterConfig*.xml) was already set before. This would lead
to a warning if a parameter occurred in multiple PI-XML files, but also
if a value was already present in the model file(s). The latter type of
warning is unnecessary as we expect parameters to be updated at runtime.
This commit makes sure that only warnings are raised for duplicates
found in the PI-XML files.