Make fix_minimized_values False for non-default solvers
Solvers like Gurobi and CPLEX cannot handle quadratic equality
constraints. The default value for fix_minimized_values being True might
therefore lead to issues if more than one minimization goal is present.

To provide backwards compatibility, we make sure that the default value
of True is preserverd for IPOPT and BONMIN. Only in the case that the
user overrides the solver will the default value be False.