Fix scale_by_problem_size for path target goals
If a path goal has a target min and/or max, the epsilon will be handled in
the normal objective (not path_objective()). We therefore have to make
sure to also divide by the number of time steps after summing an epsilon
over time if the user has enabled the "scale_by_problem_size" options.