Rename constraint_relaxation to violation_relaxation
The "constraint_relaxation" name was confusing, as the relaxation is
only indirectly applied to the eventual hard constraint (i.e. it is
scaled by the goal's function nominal as well). The term
violation_relaxation much better describes that this relaxation is
applied to the violation variables (a.k.a. epsilons).