Add alternative goal programming hard constraints
A new way of turning the soft constraints of goals into hard constraints
is added in this commit. The user can choose to use this alternative by
setting setting 'keep_soft_constraints' to True.

The new implementation keeps the soft constraints and the violation
(a.k.a. epsilon) variables around, and just constrains the objective
value that a certain priority evaluated to. Note that the results are
expected to be the same if all violations will be 0.0. Only when
violations become nonzero can the new implementation provide more room in
further priorities.

The new implementation is expected to help with infeasibilities in lower
priorities, even if a trivial goal is added. It turned out that the
current/old way of goal programming could lead to these infeasibilities,
likely because it generated entirely differently looking hard constraints.

Note that keeping the violation variables around may or may not lead to a
performance hit. Because the constraints of the problem remain entirely
the same except for one additional constraint (and the constraints of the new
priority), the solver may more readily pick up the seed solution and
actually be faster in solving the new priority.

A word of caution when using this option with (MI)QCP solvers. A
quadratic objective will turn into a quadratic constraint. In other words,
a MIQP problem can turn into a MIQCPQP problem (if the next priority also
has a quadratic objective). This is something that solvers like Gurobi and
CPLEX typically cannot handle.

Closes #1048