GoalProgramming: Add constraint relaxation option
This option was removed in deae0c3f, as its previous implementation was
somewhat of a misnomer. The "constraint_relaxation" option is now
re-added, but in a way that is consistent with its naming and intent.

The "constraint_relaxation" option should be used to add  _constant_
slack to any hard constraint generated for a goal. If a relative slack
is needed, one can use either a goal's relaxation, or the
"violation_relaxation" option.