Support vector constraints and variables
Sometimes it makes sense to make a vector symbol, and write constraints
using those vector symbols. This can be particularly interesting if many
(10,000+) constraints are added. Every call to CasADi induces a certain
overhead due to SWIG, and combining the 10,000 constraints into one, can
result in a 10,000x speedup.

Note that vector symbols are only allowed for additional path variables,
extra variables and constant inputs (i.e. symbol not part of the DAE),
and for parameters. The symbols are also not yet allowed to be
integrated states or control states.

The implementation of vector constraints is such that it does not matter
if one is using multiple single constraints, or one vector constraint
(provided the order is the same). The resulting problem will be
_exactly_ the same, and will therefore lead to _exactly_ the same
objective value.