Support vector goals
Like the commit for vector constraints, this commit allows one to write
expressions using vectors. If one would have many goals that are very
similar, one can easily combine them by vertcatting them. See the tests
in this commit for some simple examples.

This commit is primarily aimed as a performance improvement. Every call
to CasADi incurs a certain overhead due to SWIG, and writing a soft
constraint means calling CasADi 4 or 5 times _per soft constraint_. If
one has 1,000+ goals, this is no longer a neglible amount of time. If
all those goals are similar, one can combine them first by, resulting in
only a handful calls to CasADi to turn it into a soft constraint.

The implementation of vector goals is such that it does not matter if
one is using multiple single goals, or one vector goal (provided the
order is the same). The resulting problem will be _exactly_ the same,
and will therefore lead to _exactly_ the same objective value.