fix initialization error in GP mixin
A previous commit accidentally referenced the self.ensemple_size
property in the __init__ method. self.ensemble_size is not known until
after self.pre(), when [PI, CSV]Mixin read the input data. This commit
removes that reference, and makes the assumption that there are no
ensembles. This means that models using ensembles cannot call the
constraints method with with an ensemble_member parameter greater than 0
until after GP optmize has first run.

While it is unlikely that any user would encounter this issue, we should
find a real solution to this problem.