Fix indexing error in goal function value storage
For normal goals, if we would have more than one goal_function stored,
the indexing would fail because the indices were reversed. The results
would be a column vector, meaning we would have to index with `[j, :]`
instead of `[:, j]`. We transpose it into a row vector, such that we can
use the same indexing for path goals as for normal goals.

The missing broadcasting of goal function values for normal (i.e. not
path) goals is a regression introduced in 3690cfa4. Instead of relying
on np.full_like() to do broadcasting, We now explicitly make a
Timeseries object, which will do the broadcasting for us if needed.