Improve CSVLookupTableMixin and LookupTable tests
Testing of lookuptables was sloppy and not very thorough. Furthermore,
the tests treated CSVLookupTableMixin as though it is an extension of
CSVMixin, while the reality is that it is completely independant. This
commit splits the testing of CSVLookupTableMixin and the LookupTable
class into its own file, and tests both the symbolic and numerical
evaluation of LookupTable functions.