Refactor doc plotting scripts
These scripts were hard to read and not really scaling the results well.
This commit makes sure that code is legible, as many of the examples
render the plotting scripts inline. We also add the plotting scripts to
the flake8 tests, to enforce good formatting going forward.

We also now use the plotting script for the basic example and have
gotten rid of that ugly excel .png snip that was used before.