CSVMixin: expose the import/export ts base names
This addition brings the CSVMixin API more into line with the PIMIxin.
By exposing the timeseries_import_basename and
timeseries_export_basename class variables, a model can select which
filename to use for imports or exports. This feature is important when
building more complex RTC-Tools models.

We also add these variables to the class docstrings in both PIMixin and
CSVMixin, so that these variables are documented by Sphinx.

Closes #1075