CSVMixin: Output CSV files with time stamps from times() method
The data interpolation assumes that self.times() dictates the time stamps
for the output data. The time stamps in the CSV output file, however, were
taken from the input CSV file and not from the times() method. This
implicitly assumed that self.times() equates the time steps from the CSV
input file. This is not always the case, as self.times() may be overridden
by a subclass.