Disallow goal violation timeseries when keeping soft constraints
The violation timeseries was only ever set in the call to
__soft_to_hard_constraints, which is never called when
keep_soft_constraints is set.

As the goal violation is no longer guaranteed to be constant over
priorities, there is no easy way to support goal violation timeseries id
when keeping soft constraints. That is why we explicitly disallow it in
this commit.

Closes #1088