GPMixin: Disallow non-positive weight for goals with target bounds
If a goal has target bounds, a negative weight would mean maximization
of the violation variable. This makes no sense from the perspective of
goal programming, which is why it is now explicitly disallowed with a
check.

If a goal has target bounds and a weight of zero, we would basically
just be adding useless constraints. This therefore also makes no sense,
and is now explicitly disallowed.

If the goal does not have target bounds, a negative goal weight is
possible if e.g. the user wants to maximize, or flip the sign of the
goal function.