ControlTreeMixin: Improve branch allocation
Instead of picking the first ensemble member in a branch as our first
reference, we should pick the most extreme one. That way, the ordering
of the ensemble members does not have an influence on what branch they
end up on (barring cases where distance of an ensemble member to the
reference branches is equal).

Closes #1095