GPMixin: Fix scale_by_problem_size with keep_soft_constraints
The combination of these two options would lead to infeasible objective
constraints, because the size of vector symbols in the objective were
not taken into account when building the constraint.

For example, when using `scale_by_problem_size` an objective `e` with
`e` being a symbol of size 3 would lead to an (expanded) objective of
`(e[0] + e[1] + e[2]) / 3`. If this minimized to a value of 0.1, we
would end up with an objective constraint `e[0] + e[1] + e[2] <= 0.1`,
which is infeasible.