CollInt: Fix constraints on initial derivative
When history data is available at t-1 but not at t0, we would end up
using the unscaled symbol of the state, resulting in the wrong value for
the initial derivative. This was not caught in the relevant test,
because the nominal on 'w' was 1.0.

This commit fixes the bug by properly using the scaled version of the
symbol, and also changes the test such that it would have caught this
bug.

Note that the behavior was/is correct when when history data is
available at both t0 and t-1, as there is no reference to the initial
state entry in the state vector in this case (only to numeric values).