Annotate AliasDict element types in API
For dictionaries and lists we typically indicate what the type of the
elements are, but this was not yet possible for the commonly used
AliasDict class.

This commit adds type annotation support to the AliasDict class, and
also clarifies the API with the correct element types in all places that
an AliasDict is expected or returned.