ModelicaMixin: Fix reading of symbolic start attributes
Regression in 10da03cb. Only for SX symbols does a symbolic expression
turn into a NaN. For symbolic MX expressions a RuntimeError is raised.

We also have to do the Python type conversion on the _substituted_
value, not the original MX value (which will result in a RuntimeError).

Lastly, if we cannot resolve the seed value, we should continue to the
next symbol instead of still setting the seed.