ModelicaMixin: Fix symbolic parameter substitution checks
If not all parameters that occur in a symbolic expression are specified,
it can either result in a (still) symbolic expression, or NaN. We
checked the former, but did not yet check the latter.

Note that the reason that expressions result in NaN is because some of
the parameters are NaN. This in turn is caused by ModelicaMixin not
skipping parameters in the model for which a value is not set (NaN).