CollInt: Remove unnecessary resolving of bounds
The method bounds() does not return symbolic values, and resolving of
these values is therefore not needed in this class. For symbolic bounds
from a Modelica model we make sure to resolve them in ModelicaMixin.

The code that we remove/change in this commit was a remainder of the
JModelica era, where resolving of symbolic expressions with parameter
values was not done in ModelicaMixin.