CollInt: Remove unnecessary resolving of initial values
The methods history() and initial_state() do not return symbolic values,
and resolving of these values is therefore not needed in
CollocatedIntegratedOptimizationProblem. For symbolic initial state
values from the model, we make sure to resolve them right away in
ModelicaMixin.

The code that we remove/change in this commit was a remainder of the
JModelica era, where resolving with parameter values was not done in
ModelicaMixin.

Note that, depending on the completeness of the history data, initial
derivatives can be a function of the (symbolic) initial state. They are
however not not a function of e.g. symbolic parameters.