Timeseries: Fix constructor for ca.DM and support float arguments
According to the typing hints, a ca.DM should have been supported, but
it would fail with an Exception because it does not support len().

Furthermore, the code in the constructor seemed to suggest support for
floats (and integers), due to the "np.full_like" branch. The typing
hints did however not cover this, and neither did it work in practice
because of the len() call.