ModelicaMixin: Remove useless constant_inputs() implementation
If a constant input is given a default value in the Modelica model,
Pymoca will turn this assignment into an equation. All constant inputs
(and control inputs), whether they have been given a default value or
not in the Modelica model, will therefore have a value attribute equal
to NaN.

Note that this means we also cannot (easily) detect if a value has been
given to a constant input in the Modelica file. If a user gives the same
constant input another value in RTC-Tools, the system of equations could
become infeasible.