Revert "CollInt: Fix missing parameter check to include NaN"
This reverts commit 038975e752f2ff01bec36f0d0cfac951517f87da.

There are many models that have parameters that are not used. For
example, species transport in RTC-Tools Channel Flow might use a
"V_nominal" for linearization, but if there are no species this
parameter is never used.

We would rather like to check on parameters that are actually used, but
that would be an expensive check to always perform for every run. It is
probably better to put a check like that behind an optional flag for
performing such checks.