ModelicaMixin: Fix reading of ca.DM nominals
If the nominal from the model was a ca.DM instance, it would raise an
exception when calling np.isnan on it.

For performance reasons we want to avoid naively converting to float
twice, at least for the common case of the nominal being non-symbolic. To
that end, we use a similar resolving check for nominals as was already in
place for e.g seed() and bounds().