CollInt: Scale initial derivatives by nominal / dt
The scaling of initial derivatives used only the nominal, which is
different from the scaling applied to the "normal" (i.e. not "initial").
In the latter, a factor of "nominal / dt" would appear as the coefficient
in front of the state vector entries with our finite difference approach.

The wrong scaling would also lead to failing optimization when e.g. time
steps of 10 days were used, as then the scaling would be off by a factor
of almost 1E6.

This commit fixes the scaling of the initial derivatives by also dividing
by dt.

Closes #1102