GPMixin: Fix wrong IPOPT option when using minimization goals
We check if the DAE (the set of equality constraints) is linear, and tell
IPOPT if we find that is the case for performance reasons. Any user added
equality constraints are assumed to be linear as well, but for those added
by GoalProgrammingMixin for minimization goals that is not necessarily the
case. We should therefore make sure to _not_ tell IPOPT that it can assume
the equality constraints to be linear when adding such a constraint.

Note that this only applies when `fix_minimized_values` is set to True
(which is the default for IPOPT/BONMIN, but not for other solvers), and
the user is using `keep_soft_constraints` or has a minimization goal in
any but the last priority.

Closes #1103