CollInt: Explicitly disallow nominals for path and extra variables
If one specifies a nominal not equal to 1.0 for a path variable, the
bounds and the seed of this variable will be scaled wrong. For extra
variables, a nominal is simply ignored, which may be confusing in itself
when trying to properly scale a problem.

We therefore now explicitly disallow nominals for path variables by
raising an exception to prevent inconsistently applied scaling. For
extra variables we warn the user that their nominals will be ignored.