Allow nominals for path and extra variables
Up until now, we required the user to properly scale their additional
path and extra variables. In practice this meant that the user had to
multiply the variable with its nominal in every constraint where they
were using it, and also when extracting the results.

This commit adds support for nominals for path and extra variables, such
that this process is no longer required. These variables can now be
used and scaled in the same way that variables from the Modelica model
are.

Closes #1063