Make tests faster by skipping Modelica libraries
A lot of time was spent in Pymoca's transfer_model to compile the model.
That is because since d1bc9d42 we have RTC-Tools Channel Flow as a
dependency, so we would parse its entire library again for every model.

The RTC-Tools tests do not need any RTC-Tools Channel Flow components
(or other libraries for that matter), only the examples do. Therefore,
we can explicitly disable additional library folders using the
compiler_options() method to make compilation of a model much faster.